+++ templates/single-fleetcomposition.php
@@ -1,5 +1,5 @@
 <?php
-defined('ABSPATH') or die();
+\defined('ABSPATH') or die();
 ?>
 
 <header class="page-title">